requirements.txt
python-dotenv
apscheduler
python-gnupg
requests
